This change solved an issue of mine kindly review
The error was "ValueError: cannot reshape array of size 31064064 into shape (10000,32,32,3)" when I changed num_examples for test data to 10112 it becomes equal to 31064064. This worked for me, I am not an expert in this but I just want other people don't get stuck on this error. Kindly review this Thank you!